requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
requests
sentence-transformers
python-dotenv
requests
